+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 _Changes in the next release_
 
 ### Fixed
+- Fast-forward and rewind commands.
 - Automatically migrate old configuration file at startup and configuration flow, otherwise the device must be paired again.
 
 ---